[Application] Changed type of all monetary values to Decimal.

+++ SpringMvc/Models/POCO/Invoice.cs
@@ -9,7 +9,7 @@ namespace SpringMvc.Models.POCO
     {
         public virtual long Id { get; set; }
         public virtual long Counter { get; set; }
-        public virtual double TotalValue { get; set; }
+        public virtual Decimal TotalValue { get; set; }
         public virtual VatMap Vat { get; set; }
         public virtual Order Order { get; set; }
     }
